+++ src/recipe-steps/dto/update-recipe-step.dto.ts
@@ -1,5 +1,5 @@
 import { PickType } from '@nestjs/mapped-types';
-import { CreateRecipeStepDto } from './create-recipe_step.dto';
+import { CreateRecipeStepDto } from './create-recipe-step.dto';
 
 export class UpdateRecipeStepDto extends PickType(CreateRecipeStepDto, [
   'order',
